healthd: Disable the RTC wakeup timer
The timer is used here to avoid battery drained to zero. But the periodic
wakeup during suspend should not be needed as MSM PMIC already support
that by trigger interrupt to avoid the critical low battery situation.
This will save more power. And if the design is not support HW interrupt
of low battery, need to enable the timer to periodic check.

Change-Id: I040b4bc8548dc59b5e7cacb9a5a8a3557604f458

Patchset: 1
http://review.cyanogenmod.org/#/c/70170/